solana-cli: fix instruction data when requesting access (#149)
There was a bug when requesting validator access, where the signature
was verified with the correct message depending on whether backup IDs
were provided. But these backup IDs were never used when building the
instruction to be invoked on the Passport program. This PR ensures that
the correct access mode is used when building the instruction.