add sol-conversion-admin-cli (#156)
This PR includes the SOL conversion program interface.

Also associated with this change are the following `doublezero-solana`
changes:
- add `revenue-distribution relay sweep-distribution-tokens` command
- add `revenue-distribution relay finalize-distribution-rewards` command
- add deposit initialization to `revenue-distribution relay
pay-solana-validator-debt`
- clean up `revenue-distribution relay pay-solana-validator-debt`
command
- add more to `revenue-distribution fetch distribution` output

Also added a `tracing` feature to the `doublezero-solana-validator-debt`
crate, where for scheduled processes we should build with this feature
for better logging. I did it this way because of unification of
dependencies, where features become additive if one crate happened to
use `doublezero-solana-client-tools` crate with the tracing feature (so
all crates that have this as a dependency end up having this feature).
This unification was causing an issue with `doublezero-solana` output.

In a subsequent PR, the SOL conversion interface will be fleshed out
more for user interaction (e.g. buy SOL).