fix(revenue-distribution-admin): update migration command (#195)
## Summary of Changes
Companion PR to
https://github.com/doublezerofoundation/doublezero-solana/pull/87.

Whenever the Revenue Distribution program is upgraded to take this
change, admin needs to call the migration instruction after the upgrade,
which will fix the Journal state. This command needs to reflect the new
way the migration instruction must be called.

Closes https://github.com/malbeclabs/doublezero/issues/2144.

## Testing Verification
CI in https://github.com/doublezerofoundation/doublezero-solana/pull/87
forks mainnet to test this CLI command.