Add scheduler app to manage debt lifecycle processes (#183)
## Summary of Changes
This PR adds a new app called `scheduler` to manage the debt lifecycle
processes. It's written in Elixir which is a new language and adds some
risk as only Rahul and I know Elixir well. However, this is done not
simply to increase complexity but because it provides a lightweight and
reliable scheduling mechanism through Erlang's process supervision. This
could be written in Rust but it would take longer and be less reliable.
It might be worth rewriting in Rust at some point. It'd be pretty simple
as the surface area is intentionally kept as small as possible. It uses
the same entrypoint as the CLI so there aren't competing
implementations.

Rust is interfaced through a NIF (native implemented function) which is
like foreign function interface (FFI) in other languages. There's a thin
layer written in Rust that calls into the `validator-debt` crate et al.

The way this works is that there is an application
[supervisor](https://hexdocs.pm/elixir/Supervisor.html) that manages
processes. Any of the lifecycle processes will be managed through a
[GenServer](https://hexdocs.pm/elixir/1.19.2/GenServer.html) which
basically runs in a loop executing some component of the debt lifecycle
process at some interval. Intervals are configured using cron syntax.

`CHANGELOG.MD` added.

Note there is some followup work to extract some of the logic from the
CLI into a shared place, perhaps `solana-tools` but that will be done in
a subsequent PR.

Closes https://github.com/malbeclabs/doublezero/issues/2029

## Testing Verification
This has been tested running every two hours with results being posted
to slack as seen
[here](https://malbeclabs.slack.com/archives/C09LES1Q127). The app was
run from my local machine for more than an day without interruption or
errors.