fix(validator-debt): automate initialize distribution with new scheduler (#197)
## Summary of Changes
This PR adds in a worker to manage initializing distributions
automatically. It runs every two minutes. This PR also moved some of the
logic for initializing a distribution into the validator-debt worker so
that both the scheduler and the CLI have a common, minimal interface
into the underlying code.

Closes https://github.com/malbeclabs/doublezero/issues/2141.

## Testing Verification
The current distribution has already been initialized but here's
evidence of the worker calling the underlying rust code:

```
14:02:00.650 [debug] Scheduling job for execution

14:02:00.658 [debug] Task for job started on node

14:02:00.658 [debug] Execute started for job

14:02:00.760 [debug] Execution ended for job

14:02:01.328 [error] initialize_distribution: received error: "Last completed DZ epoch 52 != program's epoch 53"

14:02:01.331 [error] GenServer Scheduler.Worker.InitializeDistribution terminating
** (stop) "initialize_distribution shutting down"
Last message: {:continue, :initialize_distribution}
```