feat(validator-debt): add debt write-off logic (#225)
## Summary of Changes

This PR adds debt write-off handling prior to when the debt accountant
offchain process initializes a distribution. In order to prevent bad
debt from persisting, the accountant will first write off debt for the
distribution that will have its rewards distributable. Then it will give
its best attempt at writing off any past debt (from previous DoubleZero
epochs).

**Write-offs are explicitly disabled for now.** Once the onchain work
adds an activation epoch, there will be a subsequent PR to fetch this
activation epoch to enable the write-off feature.

Associated with this change is an update to `doublezero-solana` which
shows written-off debt in the following `revenue-distribution fetch`
commands:
- `distribution`
- `validator-debts`
- `validator-deposits`

When debt is written off, Solana validators can find how much of their
debt was marked as bad debt. At the distribution level, there will be an
indication of how much SOL is exchangeable to 2Z equating to the total
amount of debt less the uncollectible debt.

This PR also adds a `doublezero-solana-sdk`, which will aim to reduce
code repetition among crates in this repository.

And there was a bug with tracing logs, where we needed to add tracing
subscriber to the scheduler process.

Closes https://github.com/malbeclabs/doublezero/issues/2352.

## Testing Verification
Local validator test in CI should pass.