feat(solana-admin-cli/revenue-distribution): update migration command (#229)
## Summary of Changes
With [this PR] in the doublezero-solana repo, the migration instruction
will now simply reset the is-migrated bit to false because the previous
migration was already executed.

Having this command ready to execute before the program is deployed has
no impact on the CLI.

Closes https://github.com/malbeclabs/doublezero/issues/2456.

## Testing Verification
CI in [this PR] in the doublezero-solana repo should pass.

[this PR]:
https://github.com/doublezerofoundation/doublezero-solana/pull/99