validator-debt: fix local uncollectible debt tracking for write-off logic (#240)
## Summary of Changes
This change fixes the total SOL debt tracking with the local copy of the
rewards distribution and the early termination when no more debt can be
written off.

I needed to modify the `doublezero-solana-fork` CLI in order to rewind
the DZ epoch to replicate the state at the time the CI failed. Also
added a GitHub workflow to replicate the CI issue found in PR #239.

Closes https://github.com/malbeclabs/doublezero/issues/2535.

## Testing Verification
Add CI to replicate failure here:
https://github.com/doublezerofoundation/doublezero-offchain/actions/runs/20545495479/job/59015392338?pr=239.